// src/app/layout/user/models/user.ts
import { Address } from '@app/_models/address';

export class User {
  Id: number;
  BusinessName: string;
  Slogan: string;
  AddressId: number | null;
  Status: number | null;
  OperatorName: string;
  OperatorCode: string;
  AreaOfOperation: string;
  CustDcumentPath: string;
  Address: Address;
  CreatedOn: string;
}



export class Franchisee {
  Id: number;
  Name: string;
  UserId: number;
  ParentId: number | null;
  FranchiseeCode: string;
  OpeningAmount: string;
  OpeningType: number | null;
  FranciseeType: number | null;
  Description: string;
  BillNotes: string;
  Slogan: string;
  AddressId: number | null;
  Status: number | null;
  CreatedOn: string;
  Address: Address;
  User: User;
  Franchisee1: Franchisee[];
  Franchisee2: Franchisee;
}